+++ src/pages/Home/Home.js
@@ -6,6 +6,7 @@ import CarouselSlider from "../../components/CarouselSlider/CarouselSlider";
 import "./home.css";
 import GenresData from "../../Genres.json";
 import Footer from "../../components/Footer/Footer";
+import Swal from "sweetalert2";
 
 const Home = () => {
   const urlBackend = process.env.REACT_APP_URL; 
@@ -27,9 +28,16 @@ const Home = () => {
     const res = await axios.post(`${urlBackend}/movies/latest-by-genre`, bodyParameters, headersConfig)
     if(res.status === 200){
       if(res.data.estado === 401){
-        alert(res.data.mensaje);
-        localStorage.clear();
-        window.location.href="/login";
+        Swal.fire({
+          title: 'CineMagic',
+          text: res.data.mensaje,
+          icon: "success",
+          showConfirmButton: false,
+          timer: 2000
+        }).then(function() {
+          localStorage.clear();
+          window.location.href="/login";
+        })
       } else {
         setPelis(res.data.movies)
       }
